+++ src/main/java/uk/ac/ox/oxfish/fisher/equipment/gear/Gear.java
@@ -15,7 +15,7 @@ public interface Gear {
     Catch fish(
             Fisher fisher,
             SeaTile where,
-            double hoursSpentFishing,
+            int hoursSpentFishing,
             GlobalBiology modelBiology);
 
     /**
