+++ src/main/java/uk/ac/ox/oxfish/fisher/equipment/gear/RandomCatchabilityTrawl.java
@@ -42,7 +42,7 @@ public class RandomCatchabilityTrawl implements Gear
 
     @Override
     public Catch fish(
-            Fisher fisher, SeaTile where, double hoursSpentFishing, GlobalBiology modelBiology)
+            Fisher fisher, SeaTile where, int hoursSpentFishing, GlobalBiology modelBiology)
     {
         List<Species> species = modelBiology.getSpecies();
         double[] totalCatch = new double[modelBiology.getSize()];
